requirements.txt
APScheduler==3.7.0
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-environ==0.4.5
django-storages[google]==1.11.1
djangorestframework==3.12.4
google-api-core==2.0.0
google-auth==2.0.1
google-cloud-secret-manager==2.7.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.39.0
gunicorn==20.1.0
idna==2.10
libcst==0.3.20
mypy-extensions==0.4.3
packaging==21.0
proto-plus==1.19.0
protobuf==3.17.3
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
typing-inspect==0.7.1
tzlocal==2.1
urllib3==1.26.5
whitenoise==5.2.0
